requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
python-dotenv==1.0.0
gemini-webapi>=1.13.0
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
openai>=1.0.0